telegram_bot/requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
# matplotlib==3.8.0  # для графиков позже
# pandas==2.1.0       # для анализа